+++ CHANGELOG.md
@@ -30,11 +30,14 @@ per cluster.
 - Remove `amis.txt` file.
 - Remove additional EBS volume attached to the head node by default. 
 - Change NICE DCV session storage path to `/home/{UserName}`.
+- Create S3 bucket per region shared with cluster and image if custom bucket isn't specified instead creating bucket 
+per cluster.
 - Rename MasterServer to HeadNode in cli outputs.
 - Rename variable exported in the AWS Batch job environment from MASTER_IP to PCLUSTER_HEAD_NODE_IP.
 - Rename all CFN outputs from Master* to HeadNode*.
 - Rename NodeType and tags from Master to HeadNode.
 - Remove Ganglia support.
+- Add support for associating an existing Elastic IP to the head node.
 
 2.x.x
 ------